+++ app/models/playlist.rb
@@ -1,0 +1,4 @@
+class Playlist
+  include Lotus::Entity
+  self.attributes = :user_id, :provider, :external_id, :title
+end