+++ app/views/playlists/index.rb
@@ -1,6 +1,0 @@
-module GoSpotify::Views::Playlists
-  class Index
-    include GoSpotify::View
-    format :json
-  end
-end