+++ app/workers/syncer.rb
@@ -1,0 +1,7 @@
+class Syncer
+  include Sidekiq::Worker
+
+  def perform(user_id, provider, playlist_id)
+    
+  end
+end