+++ config/routes.rb
@@ -3,4 +3,6 @@ get "/auth/:provider/callback",  to: "auth#callback", as: :auth_callback
 get "/auth/failure",             to: "auth#failure",  as: :auth_failure
 get "/auth/signout",             to: "auth#signout",  as: :auth_signout
 
-resources :playlists,            only: [:show]
+namespace "/provider/:provider" do
+  resources :playlists,          only: [:index, :update]
+end